DA API Payload to Item in mempool  (#634)
* Base cleaning of da to new traits/structure
Added new da protocols and types

* DA: KZG+RS core (#632)

* Removed old kzg rs modules

* Added new kzgrs core module

* Implemented bytes_to_polynomial and tests

* Use coefficient form

* Refactor evaluations into method

* Use domain elements instead of roots of unity in tests

* Fix encoding and test

* Clippy happy

* Add comments

* Implement polynomial commitment

* Implement proof generation

* Sketch fn signature for verification

* implement proof verification

* Implemented verification and tests

* Return evaluations from bytes_to_polynomial as well

* Use modular le bytes

* Implement rs encode/decode

* Implement decoding tests

* Implement decode using lagrange

* Cleanup imports

* Reduce abstraction for certificate and vid metadata

* Allow payload to mempool as long as it converts into item

* Da Certificate verifier

* Add mock certificate for core tests

* Mempool payload verification

* Integrate mock verifiers for tx and certs

* Detach verification from cert and tx

* Seperate payload and item in mempools

* Mempools in integration tests

* Remove old cert verifier

* Network payload to item constraints in da mempool

* Update nomos-da/full-replication/src/lib.rs


* Sort attestations for cert signature

* Update nomos-da/full-replication/src/lib.rs


---------

+++ nomos-da/kzgrs/src/common.rs
@@ -110,7 +110,7 @@ mod test {
             let eval_point = DOMAIN.element(i);
             let point = poly.evaluate(&eval_point);
             // check point is the same
-            assert_eq!(evals[i as usize], point);
+            assert_eq!(evals[i], point);
             // check point bytes are the same
             assert_eq!(
                 &bytes[CHUNK_SIZE * i..CHUNK_SIZE * i + CHUNK_SIZE],
